% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantic_ants_doc.R
\docType{data}
\name{atlantic_ants}
\alias{atlantic_ants}
\title{Atlantic Ants}
\format{
A tibble with 178,976 rows and 57 variables:
\describe{
\item{id}{a number indicating the observation}
}
}
\source{
{Silva et al. 2018. ATLANTIC ANTS: a data set of ants in Atlantic Forests of South America. Ecology.}
\doi{10.1002/ecy.3580}
}
\usage{
atlantic_ants
}
\description{
ATLANTIC ANTS: a data set of ants in Atlantic Forests of South America
}
\keyword{datasets}
