% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantic_bats_doc.R
\docType{data}
\name{atlantic_bats}
\alias{atlantic_bats}
\title{Atlantic Bats}
\format{
A tibble with 21,424 rows and 37 variables:
\describe{
\item{id}{a number indicating the observation}
}
}
\source{
{Muylaert et al. 2017. ATLANTIC BATS: a data set of bat communities from the Atlantic Forests of South America. Ecology.}
\doi{10.1002/ecy.2007}
}
\usage{
atlantic_bats
}
\description{
ATLANTIC BATS: a data set of bat communities from the Atlantic Forests of South America
}
\keyword{datasets}
