% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantic_bird_traits_doc.R
\docType{data}
\name{atlantic_bird_traits}
\alias{atlantic_bird_traits}
\title{Atlantic Bird Traits}
\format{
A tibble with 72,483 rows and 83 variables:
\describe{
\item{id}{a number indicating the observation}
}
}
\source{
{Gonçalves et al. 2018. ATLANTIC BIRD TRAITS: a data set of bird morphological traits from the Atlantic forests of South America. Ecology.}
\doi{10.1002/ecy.2647}
}
\usage{
atlantic_bird_traits
}
\description{
ATLANTIC BIRD TRAITS: a data set of bird morphological traits from the Atlantic forests of South America
}
\keyword{datasets}
