% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantic_birds.R
\docType{data}
\name{atlantic_birds}
\alias{atlantic_birds}
\title{Atlantic Birds}
\format{
A tibble with 72,483 rows and 83 variables:
\describe{
\item{id}{a number indicating the observation}
}
}
\source{
{Hasui et al. 2017. ATLANTIC BIRDS: a data set of bird species from the Brazilian Atlantic Forest. Ecology, 99(2), 497-497.}
\doi{10.1002/ecy.2119}
}
\usage{
atlantic_birds
}
\description{
ATLANTIC BIRDS: a data set of bird species from the Brazilian Atlantic Forest
}
\keyword{datasets}
