% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantic_butterflies_doc.R
\docType{data}
\name{atlantic_butterflies}
\alias{atlantic_butterflies}
\title{Atlantic butterflies}
\format{
A tibble with 7,062 rows and 25 variables:
\describe{
\item{id}{a number indicating the observation}
}
}
\source{
{Santos et al. 2018. Atlantic butterflies: a data set of fruit-feeding butterfly communities from the Atlantic forests. Ecology.}
\doi{10.1002/ecy.2507}
}
\usage{
atlantic_butterflies
}
\description{
Atlantic butterflies: a data set of fruit-feeding butterfly communities from the Atlantic forests
}
\keyword{datasets}
