% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantic_camera_trap_mammals_doc.R
\docType{data}
\name{atlantic_camera_trap_mammals}
\alias{atlantic_camera_trap_mammals}
\title{Atlantic Camera Trap Mammals}
\format{
A tibble with 13,940 rows and 37 variables:
\describe{
\item{id}{a number indicating the observation}
}
}
\source{
{Franceschi et al. 2024. Camera trap surveys of Atlantic Forest mammals: A data set for analyses considering imperfect detection (2004–2020). Ecology, 105(5), e4298.}
\doi{https://doi.org/10.1002/ecy.4298}
}
\usage{
atlantic_camera_trap_mammals
}
\description{
Camera trap surveys of Atlantic Forest mammals: A data set for analyses considering imperfect detection (2004–2020)
}
\keyword{datasets}
