% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantic_camtraps_doc.R
\docType{data}
\name{atlantic_camtraps}
\alias{atlantic_camtraps}
\title{Atlantic Camtraps}
\format{
A tibble with 13,940 rows and 37 variables:
\describe{
\item{id}{a number indicating the observation}
}
}
\source{
{Lima et al. 2017. ATLANTIC-CAMTRAPS: a dataset of medium and large terrestrial mammal communities in the Atlantic Forest of South America. Ecology, 98(11), 2979-2979.}
\doi{https://doi.org/10.1002/ecy.1998}
}
\usage{
atlantic_camtraps
}
\description{
ATLANTIC-CAMTRAPS: a dataset of medium and large terrestrial mammal communities in the Atlantic Forest of South America
}
\keyword{datasets}
