% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantic_epiphytes_doc.R
\docType{data}
\name{atlantic_epiphytes}
\alias{atlantic_epiphytes}
\title{Atlantic Epiphytes}
\format{
A tibble with 100,305 rows and 44 variables:
\describe{
\item{id}{a number indicating the observation}
}
}
\source{
{Ramos et al. 2019. ATLANTIC EPIPHYTES: a data set of vascular and non-vascular epiphyte plants and lichens from the Atlantic Forest. Ecology.}
\doi{10.1002/ecy.2541}
}
\usage{
atlantic_epiphytes
}
\description{
ATLANTIC EPIPHYTES: a data set of vascular and non-vascular epiphyte plants and lichens from the Atlantic Forest
}
\keyword{datasets}
