% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantic_flower_invertebrate_interactions_doc.R
\docType{data}
\name{atlantic_flower_invertebrate_interactions}
\alias{atlantic_flower_invertebrate_interactions}
\title{Atlantic Flower-Invertebrate Interactions}
\format{
A tibble with 18,447 rows and 62 variables:
\describe{
\item{id}{a number indicating the observation}
}
}
\source{
{Boscolo et al. 2022. Atlantic flower–invertebrate interactions: A data set of occurrence and frequency of floral visits. Ecology.}
\doi{10.1002/ecy.3900}
}
\usage{
atlantic_flower_invertebrate_interactions
}
\description{
Atlantic flower–invertebrate interactions: A data set of occurrence and frequency of floral visits
}
\keyword{datasets}
