% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantic_frugivory_doc.R
\docType{data}
\name{atlantic_frugivory}
\alias{atlantic_frugivory}
\title{Atlantic Frugivory}
\format{
A tibble with 8,320 rows and 36 variables:
\describe{
\item{id}{a number indicating the observation}
}
}
\source{
{Bello et al. 2017. Atlantic frugivory: a plant–frugivore interaction data set for the Atlantic Forest. Ecology.}
\doi{10.1002/ecy.1818}
}
\usage{
atlantic_frugivory
}
\description{
Atlantic frugivory: a plant–frugivore interaction data set for the Atlantic Forest
}
\keyword{datasets}
