% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantic_mammal_traits_doc.R
\docType{data}
\name{atlantic_mammal_traits}
\alias{atlantic_mammal_traits}
\title{Atlantic Mammal Traits}
\format{
A tibble with 39,850 rows and 19 variables:
\describe{
\item{id}{a number indicating the observation}
}
}
\source{
{Gonçalves et al. 2018. ATLANTIC MAMMAL TRAITS: a data set of morphological traits of mammals in the Atlantic Forest of South America. Ecology.}
\doi{10.1002/ecy.2106}
}
\usage{
atlantic_mammal_traits
}
\description{
ATLANTIC MAMMAL TRAITS: a data set of morphological traits of mammals in the Atlantic Forest of South America
}
\keyword{datasets}
