% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantic_mammals_doc.R
\docType{data}
\name{atlantic_mammals}
\alias{atlantic_mammals}
\title{Atlantic Mammals}
\format{
A tibble with 4,680 rows and 40 variables:
\describe{
\item{id}{a number indicating the observation}
}
}
\source{
{Souza et al. 2019. ATLANTIC MAMMALS: a data set of assemblages of medium- and large-sized mammals of the Atlantic Forest of South America. Ecology.}
\doi{10.1002/ecy.2785}
}
\usage{
atlantic_mammals
}
\description{
ATLANTIC MAMMALS: a data set of assemblages of medium- and large-sized mammals of the Atlantic Forest of South America
}
\keyword{datasets}
