% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantic_nonvolant_mammals_doc.R
\docType{data}
\name{atlantic_nonvolant_mammals}
\alias{atlantic_nonvolant_mammals}
\title{Atlantic Non-volant mammals}
\format{
A tibble with 15,793 rows and 65 variables:
\describe{
\item{id}{a number indicating the observation}
}
}
\source{
{Gonçalves et al. 2017. Non-volant mammals from the Upper Paraná River Basin: a data set from a critical region for conservation in Brazil. Ecology, 99(2), 499-499.}
\doi{https://doi.org/10.1002/ecy.2107}
}
\usage{
atlantic_nonvolant_mammals
}
\description{
Non-volant mammals from the Upper Paraná River Basin: a data set from a critical region for conservation in Brazil
}
\keyword{datasets}
