% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantic_pollination_doc.R
\docType{data}
\name{atlantic_pollination}
\alias{atlantic_pollination}
\title{Atlantic Pollination}
\format{
A tibble with 3,762 rows and 20 variables:
\describe{
\item{id}{}
}
}
\source{
{Iamara-Nogueira et al. ATLANTIC POLLINATION: a data set of flowers and interaction with nectar-feeding vertebrates from the Atlantic Forest. Ecology.}
\doi{10.1002/ecy.3595}
}
\usage{
atlantic_pollination
}
\description{
ATLANTIC POLLINATION: a data set of flowers and interaction with nectar-feeding vertebrates from the Atlantic Forest
}
\keyword{datasets}
