% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantic_primates_doc.R
\docType{data}
\name{atlantic_primates}
\alias{atlantic_primates}
\title{Atlantic Primates}
\format{
A tibble with 2,061 rows and 31 variables:
\describe{
\item{id}{a number indicating the observation}
}
}
\source{
{Culot et al. 2018. ATLANTIC-PRIMATES: a dataset of communities and occurrences of primates in the Atlantic Forests of South America. Ecology, 100(1), e02525.}
\doi{https://doi.org/10.1002/ecy.2525}
}
\usage{
atlantic_primates
}
\description{
ATLANTIC-PRIMATES: a dataset of communities and occurrences of primates in the Atlantic Forests of South America
}
\keyword{datasets}
