% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantic_small_mammal_doc.R
\docType{data}
\name{atlantic_small_mammal}
\alias{atlantic_small_mammal}
\title{Atlantic Small Mammal}
\format{
A tibble with 4,922 rows and 37 variables:
\describe{
\item{id}{a number indicating the observation}
}
}
\source{
{Bovendorp et al. 2017. Atlantic small-mammal: a dataset of communities of rodents and marsupials of the Atlantic forests of South America. Ecology, 98(8), 2226-2226.}
\doi{https://doi.org/10.1002/ecy.1893}
}
\usage{
atlantic_small_mammal
}
\description{
Atlantic small-mammal: a dataset of communities of rodents and marsupials of the Atlantic forests of South America
}
\keyword{datasets}
