% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantic_small_mammal_abundance_doc.R
\docType{data}
\name{atlantic_small_mammal_abundance}
\alias{atlantic_small_mammal_abundance}
\title{Atlantic Small Mammal Abundance}
\format{
A tibble with 1,902 rows and 29 variables:
\describe{
\item{id}{a number indicating the observation}
}
}
\source{
{Figueiredo et al. 2017. Abundance of small mammals in the Atlantic Forest (ASMAF): a data set for analyzing tropical community patterns. Ecology, 98(11), 2981-2981.}
\doi{https://doi.org/10.1002/ecy.2005}
}
\usage{
atlantic_small_mammal_abundance
}
\description{
Abundance of small mammals in the Atlantic Forest (ASMAF): a data set for analyzing tropical community patterns
}
\keyword{datasets}
