% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantic_spatial_doc.R
\docType{data}
\name{atlantic_spatial}
\alias{atlantic_spatial}
\title{Atlantic Spatial}
\format{
A tibble with 500 rows and 10 variables:
\describe{
\item{metric}{metric}
\item{description}{description}
\item{values}{values}
\item{scale}{scale}
\item{edge_depth}{edge_depth}
\item{gap_crossing}{gap_crossing}
\item{resolution}{resolution}
\item{file}{file}
\item{link}{link}
}
}
\source{
{Vancine et al. ATLANTIC SPATIAL: a dataset of landscape, topographic, hydrologic and anthropogenic metrics for the Atlantic Forests of South America. Ecology.}
\doi{in prep.}
}
\usage{
atlantic_spatial
}
\description{
ATLANTIC SPATIAL: a dataset of landscape, topographic, hydrologic and anthropogenic metrics for the Atlantic Forests of South America
}
\keyword{datasets}
