% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantic_spatial_download.R
\name{atlantic_spatial_download}
\alias{atlantic_spatial_download}
\title{Atlantic Spatial}
\usage{
atlantic_spatial_download(metric, scale, edge_depth, gap_crossing, path = ".")
}
\arguments{
\item{metric}{\cr}

\item{scale}{\cr}

\item{edge_depth}{\cr}

\item{gap_crossing}{\cr}

\item{path}{\cr}
}
\description{
ATLANTIC SPATIAL: a dataset of landscape, topographic, hydrologic and anthropogenic metrics for the Atlantic Forests of South America
}
\examples{
library(atlanticr)

# read table with metrics
atlantic_spatial

# import data
atlanticr::atlantic_spatial_download(path = ".")

# delete
unlink("001_atlantic_spatial_all_classes.tif")
}
