% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantic_amphibians.R,
%   R/atlantic_bird_traits.R, R/atlantic_birds.R, R/atlantic_butterflies.R,
%   R/atlantic_camtraps.R, R/atlantic_epiphytes.R, R/atlantic_frugivory.R,
%   R/atlantic_mammal_traits.R, R/atlantic_mammals.R,
%   R/atlantic_nonvolant_mammals.R, R/atlantic_pollination.R,
%   R/atlantic_primates.R, R/atlantic_small_mammal.R,
%   R/atlantic_small_mammal_abundance.R, R/atlantic_spatial.R
\name{atlantic_amphibians}
\alias{atlantic_amphibians}
\title{Atlantic Amphibians}
\usage{
atlantic_amphibians(type)

atlantic_amphibians(type)

atlantic_amphibians(type)

atlantic_amphibians(type)

atlantic_amphibians(type)

atlantic_amphibians(type)

atlantic_amphibians(type)

atlantic_amphibians(type)

atlantic_amphibians(type)

atlantic_amphibians(type)

atlantic_amphibians(type)

atlantic_amphibians(type)

atlantic_amphibians(type)

atlantic_amphibians(type)

atlantic_amphibians(type)
}
\arguments{
\item{type}{data format, "long" or "wide".}
}
\description{
ATLANTIC AMPHIBIANS: a data set of amphibian communities from the Atlantic Forests of South America

ATLANTIC AMPHIBIANS: a data set of amphibian communities from the Atlantic Forests of South America

ATLANTIC AMPHIBIANS: a data set of amphibian communities from the Atlantic Forests of South America

ATLANTIC AMPHIBIANS: a data set of amphibian communities from the Atlantic Forests of South America

ATLANTIC AMPHIBIANS: a data set of amphibian communities from the Atlantic Forests of South America

ATLANTIC AMPHIBIANS: a data set of amphibian communities from the Atlantic Forests of South America

ATLANTIC AMPHIBIANS: a data set of amphibian communities from the Atlantic Forests of South America

ATLANTIC AMPHIBIANS: a data set of amphibian communities from the Atlantic Forests of South America

ATLANTIC AMPHIBIANS: a data set of amphibian communities from the Atlantic Forests of South America

ATLANTIC AMPHIBIANS: a data set of amphibian communities from the Atlantic Forests of South America

ATLANTIC AMPHIBIANS: a data set of amphibian communities from the Atlantic Forests of South America

ATLANTIC AMPHIBIANS: a data set of amphibian communities from the Atlantic Forests of South America

ATLANTIC AMPHIBIANS: a data set of amphibian communities from the Atlantic Forests of South America

ATLANTIC AMPHIBIANS: a data set of amphibian communities from the Atlantic Forests of South America

ATLANTIC AMPHIBIANS: a data set of amphibian communities from the Atlantic Forests of South America
}
