% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantic_amphibians.R
\name{atlantic_amphibians}
\alias{atlantic_amphibians}
\title{Atlantic Amphibians}
\usage{
atlantic_amphibians(type)
}
\arguments{
\item{type}{data format, "long" or "wide".}
}
\description{
ATLANTIC AMPHIBIANS: a data set of amphibian communities from the Atlantic Forests of South America
}
