% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantic_amphibians_doc.R
\docType{data}
\name{atlantic_amphibians}
\alias{atlantic_amphibians}
\title{Atlantic Amphibians}
\format{
A tibble with 15,793 rows and 65 variables:
\describe{
\item{id}{a number indicating the observation}
}
}
\source{
{Vancine et al. 2018. ATLANTIC AMPHIBIANS: a data set of amphibian communities from the Atlantic Forests of South America. Ecology, 99(7), 1692-1692.}
\doi{https://doi.org/10.1002/ecy.2392}
}
\usage{
atlantic_amphibians
}
\description{
ATLANTIC AMPHIBIANS: a data set of amphibian communities from the Atlantic Forests of South America
}
\keyword{datasets}
