% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantic_ants.R
\name{atlantic_ants}
\alias{atlantic_ants}
\title{Atlantic Ants}
\usage{
atlantic_ants(type)
}
\arguments{
\item{type}{data format, "long" or "wide".}
}
\description{
ATLANTIC ANTS: a data set of ants in Atlantic Forests of South America
}
