% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantic_bats.R
\name{atlantic_bats}
\alias{atlantic_bats}
\title{Atlantic Bats}
\usage{
atlantic_bats(type)
}
\arguments{
\item{type}{data format, "long" or "wide".}
}
\description{
ATLANTIC BATS: a data set of bat communities from the Atlantic Forests of South America
}
